--- Microsoft.Azure.Cosmos.Samples/Usage/CFPullModelAllVersionsAndDeletesMode/Models/AllVersionsAndDeletesCFResponse.cs
@@ -1,16 +1,0 @@
-﻿namespace Cosmos.Samples.CFPullModelAllVersionsAndDeletesMode.Models
-{
-    using Newtonsoft.Json;
-
-    internal class AllVersionsAndDeletesCFResponse
-    {
-        [JsonProperty("current")]
-        public Item Current { get; set; }
-
-        [JsonProperty("previous")]
-        public Item Previous { get; set; }
-
-        [JsonProperty("metadata")]
-        public Metadata Metadata { get; set; }
-    }
-}
