--- Microsoft.Azure.Cosmos.Samples/Usage/CFPullModelAllVersionsAndDeletesMode/Models/Metadata.cs
@@ -1,14 +1,0 @@
-﻿namespace Cosmos.Samples.CFPullModelAllVersionsAndDeletesMode.Models
-{
-    using System;
-    using Newtonsoft.Json;
-
-    internal class Metadata
-    {
-        [JsonProperty("operationType")]
-        public string OperationType { get; set; }
-
-        [JsonProperty("timeToLiveExpired")]
-        public Boolean TimeToLiveExpired { get; set; }
-    }
-}
