--- Microsoft.Azure.Cosmos.Samples/Usage/CFPullModelLatestVersionMode/Models/Item.cs
@@ -1,14 +1,0 @@
-﻿namespace Cosmos.Samples.CFPullModelLatestVersionMode.Models
-{
-    using Newtonsoft.Json;
-
-    internal class Item
-    {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        public double Value { get; set; }
-
-        public string Pk { get; set; }
-    }
-}
